Initialize phase masks using Field dx/spectrum
Previously, the phase masks accepted only a PRNGKey and a shape when
initialized as parameters. This matches the pattern of nn.initializers, but
requires the user to precalculate the spacing which can be awkward when trying
to match the spacing of a system. Now, properties of the Field are passed to
the phase mask initializers, as requested in issue #30.