Allow PRNGKey argument to be injected in trainable
The trainable function accepts initializer functions, but
Flax expects the first argument to initializer functions to be
a PRNGKey. Since a lot of the functions we use in Chromatix don't
need a PRNGKey, we should have this option to automatically ignore
the PRNGKey argument passed by Chromatix when calling a function
from Chromatix to initialize an optical element, e.g. a phase mask.