Make Field vmappable
Fields were not vmappable because of the dx, spectrum, and spectral_density
attributes. These should not have been treated as PyTree nodes, which would
require them to have the same batch dimension sizes as the complex field
attribute u. Instead, they should be broadcasted across the batch dimensions.
We can fix this by specifying to flax.struct that these fields should not count
as PyTree nodes.

However, this also means that in order for these attributes to have the
appropriate broadcasted shapes under a vmap, they must be dynamically reshaped
based on the current ndim of the Field. This means that we have had to prefix
these attributes with an underscore (to indicate they are no longer for public
use) and then add dx, spectrum, and spectral_density properties which get
reshaped dynamically to support vmap.