Avoid creating JAX Arrays on import
We set some default values to be JAX arrays throughout the library (when
setting default values for kykx for propagation kernels, etc.) which would
allocate GPU memory on import of certain modules. To prevent this behavior,
we can simply use a tuple (0.0, 0.0) instead. Because these values are passed
through jnp.atleast_1d anyway, it doesn't matter and we avoid the accidental
Array creation.